At the moment only the maximum number of frames to be acquired can be configured for the Hameg/Rohde&Schwarz HMO mixed-signal oscilloscope series driver (hameg-hmo).
This patch adds support to configure the number of samples to acquire
in both analog and digital (logic) mode.

The invalid product that was removed since v2 and the missing product
that was added in v2 are now treated in the following separate patch:

hameg-hmo: Remove an invalid product model and add a valid product model

which should be applied before this patch.

The extra functionality to support 16 digital channels on the HMO3XX2
models that was previously introduced in version 2 of this patch, is
now provided by the following separate patch:

hameg-hmo: support all 16 instead of only 8 digital channels on HMO3XX2

which should be applied after the patch mentioned above and before
this patch.

The fourth version (v4) preserves the bug fix introduced in the third
version (segmentation fault when trying to acquire digital data using
the limit on the number of frames instead of the limit on the number of
samples).

This fifth version (v5) simply corrects the patch ordering comment above
and removes an unneeded but totally harmless diff line.

Signed-off-by: Guido Trentalancia <guido@trentalancia.com>
---
 src/hardware/hameg-hmo/api.c      |    8 ++++++++
 src/hardware/hameg-hmo/protocol.c |   20 +++++++++++++++++---
 src/hardware/hameg-hmo/protocol.h |    2 ++
 3 files changed, 27 insertions(+), 3 deletions(-)