Introduce the support in the "rohde-schwarz-hameg" driver for the RTB2000 and RTM3000 mixed-signal oscilloscope series.
According to the latest available version of the manual, they
both use a slightly different dialect than currently supported
models, in particular when it comes to the POD (logic channel
groups) handling.

I do not have any of the above models available for testing
therefore I do not know whether or not they also understand
the existing dialect. In doubt, the new official dialect is
implemented by this patch.

Signed-off-by: Guido Trentalancia <guido@trentalancia.com>
---
 src/hardware/rohde-schwarz-hameg/api.c      |   71 +++++-
 src/hardware/rohde-schwarz-hameg/protocol.c |  287 ++++++++++++++++++++++++++--
 src/hardware/rohde-schwarz-hameg/protocol.h |    1
 3 files changed, 332 insertions(+), 27 deletions(-)