Introduce the support in the "rohde-schwarz-hameg" driver for the RTA4000 mixed-signal oscilloscope series.
According to the latest available version of the manual, as
for the RTB2000 and RTM3000 series, the RTA4000 series also
uses a slightly different dialect than other previously
supported models, in particular when it comes to the POD
(logic channel groups) handling.

I do not have such model available for testing therefore, as
for the RTB2000 and RTM3000 support recently introduced, I do
not know whether or not the RTA4000 also understands the
existing dialect. In doubt, the new official dialect is
implemented by this patch.

Signed-off-by: Guido Trentalancia <guido@trentalancia.com>
---
 src/hardware/rohde-schwarz-hameg/protocol.c |   44 +++++++++++++++++++++++++++-
 1 file changed, 43 insertions(+), 1 deletion(-)