Rename the improper "vdiv" setting to "vscale".
The vertical scale setting is improperly named "vdiv" and as
such it creates a great deal of confusion especially in new
users and developers because it is easily mistaken with
"num_vdiv" which is the number of vertical divisions !

Appropriate measures are also introduced so that the vertical
scale unit is correctly printed as "V/div" instead of "V" (a
very small companion patch has been created for sigrok-cli).

It is up to individual driver maintainers to fix the local
variable names and comments (such source-cosmetic fixes are
provided here only for the "rohde-schwarz-hameg" driver).

Signed-off-by: Guido Trentalancia <guido@trentalancia.com>
---
 bindings/cxx/ConfigKey_methods.cpp          |    4 +
 include/libsigrok/libsigrok.h               |    7 +--
 include/libsigrok/proto.h                   |    2
 src/hardware/hantek-6xxx/api.c              |    8 +--
 src/hardware/hantek-dso/api.c               |    8 +--
 src/hardware/hung-chang-dso-2100/api.c      |   10 ++--
 src/hardware/lecroy-xstream/api.c           |    8 +--
 src/hardware/lecroy-xstream/protocol.c      |    4 -
 src/hardware/rigol-ds/api.c                 |    8 +--
 src/hardware/rohde-schwarz-hameg/api.c      |   18 ++++----
 src/hardware/rohde-schwarz-hameg/protocol.c |   58 ++++++++++++++--------------
 src/hardware/rohde-schwarz-hameg/protocol.h |    6 +-
 src/hardware/siglent-sds/api.c              |    8 +--
 src/hardware/yokogawa-dlm/api.c             |    8 +--
 src/hardware/yokogawa-dlm/protocol.c        |    4 -
 src/hwdriver.c                              |    5 +-
 src/strutil.c                               |   54 +++++++++++++++++++++++++-
 tests/strutil.c                             |   25 ++++++++++++
 18 files changed, 165 insertions(+), 80 deletions(-)