Use the Arithmetics Type option instead of the Acquisition Type in the Rohde&Schwarz/Hameg driver.
The SCPI command currently described in the SCPI Programmer's
Manual as capable of setting the Acquisition Type is wrong
according to the Rohde&Schwarz Customer Service (RS# 649100).

An alternative setting (and command) described as Arithmetics
Type should be used instead, because otherwise the device
doesn't properly enter the Average and/or Envelope modes.

As a result, the Roll mode is left out of the available
options and will be introduced as a separate new feature
with a separate dedicated SCPI command.

See also:

commit 468748825c0b33bbf5b9f28a950a6c4d8a22fc31
commit 41cd818f793a796c49f0a6f1cd7ee528f2cd4db3

Signed-off-by: Guido Trentalancia <guido@trentalancia.com>
---
 include/libsigrok/libsigrok.h                 |    3 +
 src/hardware/rohde-schwarz-hameg/api.c        |   34 +++++++++++++++
 src/hardware/rohde-schwarz-hameg/model_desc.h |   58 ++++++++++++++++++++++++++
 src/hardware/rohde-schwarz-hameg/protocol.c   |    9 ++++
 src/hardware/rohde-schwarz-hameg/protocol.h   |    4 +
 src/hwdriver.c                                |    2
 src/scpi.h                                    |    2
 7 files changed, 112 insertions(+)