Use the Arithmetics Type option instead of the Acquisition Type in the Rohde&Schwarz/Hameg driver.
The SCPI command currently described in the SCPI Programmer's
Manual as capable of setting the Acquisition Type is wrong
according to the Rohde&Schwarz Customer Service (RS# 649100).

An alternative setting (and command) described as Arithmetics
Type should be used instead, because otherwise the device
doesn't properly enter the Average and/or Envelope modes.

As a result, the Roll mode is left out of the available
options and will be introduced as a separate new feature
with a separate dedicated SCPI command.