Avoid cleaning up the output modules more than once.
Currently, if the analog module is explicitly requested
in sigrok-cli, the program attempts to clean it twice,
thus resulting in g_variant_unref() assertion failures
at runtime.

This patch makes sure that the analog output module is
not cleaned up more than once. The same check is applied
to the other output modules for consistency.

Signed-off-by: Guido Trentalancia <guido@trentalancia.com>
---
 src/output/analog.c       |   21 +++++++++++++++++--
 src/output/ascii.c        |   30 ++++++++++++++++++++++------
 src/output/bits.c         |   25 ++++++++++++++++++-----
 src/output/chronovu_la8.c |   11 ++++++++++
 src/output/csv.c          |   49 +++++++++++++++++++++++++++++++---------------
 src/output/hex.c          |   28 +++++++++++++++++++++-----
 src/output/ols.c          |   11 ++++++++++
 src/output/srzip.c        |   19 +++++++++++++++++
 src/output/vcd.c          |   11 ++++++++++
 src/output/wav.c          |   20 +++++++++++++++++-
 10 files changed, 191 insertions(+), 34 deletions(-)